requirements.txt
pytorch-lightning==2.0.0
wandb==0.14.0
gensim==4.3.1
pandas==1.5.3
scikit-learn==1.2.2
tensorflow==2.12.0
bnlp_toolkit==3.0.0
bnunicodenormalizer==0.1.6
transformers[sentencepiece]
git+https://github.com/csebuetnlp/normalizer
gdown